Refactor error handling in AnnotationParser task
Do all error handling in the AnnotationParser task via exceptions,
as recommended for a Cytoscape task, and get rid of parsing/error status
tracking via "LOADCORRECT" and error string method return values.

In SettingsPanelActionListener, use a TaskObserver on the
AnnotationParser task to detect if the task failed (i.e., threw an
exception) and show the error/exception message to the user using a
JOptionPane error dialog.